dashboard/backend/requirements.txt
# SocialScope Dashboard Backend Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
firebase-admin>=6.4.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
